+++ kurtosis-module/impl/near_module.ts
@@ -148,9 +148,4 @@ export class NearModule implements ExecutableKurtosisModule {
     // ====================================================================================================
 
 
-
-
-
-
-
 }